Three changes made.
First, a test to ensure the new functionality in US-18 executes as expected -> if a pet is on an 'Approved' application (all pet apps also approved), if that pet is currently on any 'Pending' applications, it should only display a 'Reject' button next to it on that application's admin app show page.

Add a 'Pending' test for any application that has at least one 'Approved' pet application and one 'Pending' application. The general app should display 'Pending'.

Modify status of all applications and pet applications to 'Pending', making the tests easier to perform.